revert simpler TimedRwLock
