+++ src/lib.rs
@@ -9,7 +9,7 @@ mod ethereum_subscriber;
 mod identity_committer;
 pub mod identity_tree;
 pub mod server;
-mod timed_read_progress_lock;
+mod timed_rw_lock;
 mod utils;
 
 use crate::app::App;
