+++ src/server.rs
@@ -109,7 +109,7 @@ pub enum Error {
     #[error(transparent)]
     Elapsed(#[from] tokio::time::error::Elapsed),
     #[error(transparent)]
-    LockTimeout(#[from] crate::timed_read_progress_lock::Error),
+    LockTimeout(#[from] crate::timed_rw_lock::Error),
     #[error(transparent)]
     Other(#[from] EyreError),
 }
